+++ intelmq/lib/bot.py
@@ -33,8 +33,14 @@ class Bot(object):
     __current_message = None
     __message_counter = 0
     __message_counter_start = None
+
     # Bot is capable of SIGHUP delaying
     sighup_delay = True
+    # From the runtime configuration
+    description = None
+    group = None
+    module = None
+    name = None
 
     def __init__(self, bot_id: str):
         self.__log_buffer = []
@@ -214,6 +220,10 @@ class Bot(object):
                             self.parameters.error_max_retries):
 
                         if error_on_message:
+
+                            if self.parameters.error_forward_message and self.group == "Expert":
+                                self.logger.info("Forwarding message to output queue.")
+                                self.send_message(self.__current_message)
 
                             if self.parameters.error_dump_message:
                                 error_traceback = traceback.format_exception(*error_on_message)
@@ -429,6 +439,9 @@ class Bot(object):
 
         setattr(self.parameters, 'logging_path', DEFAULT_LOGGING_PATH)
 
+        # backwards compatibility
+        self.parameters.error_forward_message = getattr(self.parameters, 'error_forward_message', None)
+
         for option, value in config.items():
             setattr(self.parameters, option, value)
             self.__log_configuration_parameter("defaults", option, value)
@@ -440,7 +453,7 @@ class Bot(object):
         config = utils.load_configuration(RUNTIME_CONF_FILE)
         reinitialize_logging = False
 
-        if self.__bot_id in list(config.keys()):
+        if self.__bot_id in config:
             params = config[self.__bot_id]
             self.run_mode = params.get('run_mode', 'continuous')
             for option, value in params['parameters'].items():
@@ -448,6 +461,11 @@ class Bot(object):
                 self.__log_configuration_parameter("runtime", option, value)
                 if option.startswith('logging_'):
                     reinitialize_logging = True
+
+            self.description = params.get('description')
+            self.group = params.get('group')
+            self.module = params.get('module')
+            self.name = params.get('name')
 
         if reinitialize_logging:
             self.logger.handlers = []  # remove all existing handlers
@@ -575,6 +593,7 @@ class ParserBot(Bot):
             self.logger.error('ParserBot can\'t be started itself. '
                               'Possible Misconfiguration.')
             self.stop()
+        self.group = 'Parser'
 
     def parse_csv(self, report: dict):
         """
@@ -737,6 +756,7 @@ class CollectorBot(Bot):
             self.logger.error('CollectorBot can\'t be started itself. '
                               'Possible Misconfiguration.')
             self.stop()
+        self.group = 'Collector'
 
     def __filter_empty_report(self, message: dict):
         if 'raw' not in message: