+++ CHANGELOG.md
@@ -35,6 +35,7 @@ CHANGELOG
   This allows shorter code in the bots, as an 'overwrite' configuration parameter can be directly passed to the function.
 - Bots can specify a static method `check(parameters)` which can perform individual checks specific to the bot.
   These functions will be called by `intelmqctl check` if the bot is configured with the given parameters
+- lib/bot: top level bot parameters (description, group, module, name) are exposed as members of the class.
 - Add `RewindableFileHandle` to utils making handling of CSV files more easy (optionally)
 
 ### Bots